there is now a DAO for the statistics service. for start it has a method that counts nomenclatural references much faster than the further using of reference DAO.

+++ cdmlib-services/src/main/java/eu/etaxonomy/cdm/api/service/statistics/Statistics.java
@@ -64,7 +64,7 @@ public class Statistics {
 		return countMap;
 	}
 
-	public void addCount(StatisticsTypeEnum type, Integer number) {
+	public void addCount(StatisticsTypeEnum type, Long number) {
 		this.countMap.put(type.getLabel(), number);
 	}
